chore: basic snap and general services split
